Address CodeRabbit review comments on UBSan PR
- slang-check-expr: use signed comparisons for <, <=, >, >= (uint64_t cast
  broke e.g. (-1 < 0)); split shift handling with negative count guard
- slang-ir-sccp: mask shift counts by IR type bit-width via maybeGetIntTypeWidth
- slang-dictionary.h: add explicit static_cast<int> for narrowing
- slang-hash.h: tighten static_assert to match std::make_unsigned_t requirements

+++ source/core/slang-dictionary.h
@@ -473,7 +473,7 @@ private:
     inline int getHashPos(T& key) const
     {
         const unsigned int hash = (unsigned int)getHashCode(key);
-        return ((hash * 2654435761UL) % m_bucketCountMinusOne);
+        return static_cast<int>((hash * 2654435761UL) % m_bucketCountMinusOne);
     }
     template<typename T>
     FindPositionResult findPosition(const T& key) const
